+++ docs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [0.4.0] - 2025-12-17
+* State Collections
+
+## [0.3.0] - 2025-12-16
+* Builder Pattern 
+* Segments
+* complete API refactoring
+
+## [0.2.0] - 2025-12-11
+
+* Point2D for States
+
 ## [0.1.0] - 2025-12-08
 
 * Initial release